+++ src/app/Site/dashboard/sidebard.service.ts
@@ -1,19 +1,16 @@
-import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { SideBarItemModel } from "./sidebar-item.model";
+import { AngularFireDatabase } from '@angular/fire/compat/database';
 
 @Injectable({
     providedIn: 'root'
 })
 export class DashboardsService {
-    private baseUrl:string = "https://human-benchmark-d54ce-default-rtdb.firebaseio.com/";
-    private dashboardsEndpoint:string = "sidebar.json";
-
-    constructor(private http:HttpClient) {
+    constructor(private db:AngularFireDatabase) {
 
     }
 
     public getDashboards() {
-        return this.http.get<SideBarItemModel []>(this.baseUrl + this.dashboardsEndpoint);    
+          return this.db.list<SideBarItemModel>("sidebar").valueChanges();
     }
 }