+++ src/environments/environment.prod.ts
@@ -1,3 +1,13 @@
 export const environment = {
+  firebase: {
+    projectId: 'human-benchmark-d54ce',
+    appId: '1:751337916116:web:35fc1af11045f3f0ba9d41',
+    databaseURL: 'https://human-benchmark-d54ce-default-rtdb.firebaseio.com',
+    storageBucket: 'human-benchmark-d54ce.appspot.com',
+    apiKey: 'AIzaSyAHsOQlaf2nZjNvj-yZXZkMJY8kOOW2Y1Y',
+    authDomain: 'human-benchmark-d54ce.firebaseapp.com',
+    messagingSenderId: '751337916116',
+    measurementId: 'G-L5TKVKE5KY',
+  },
   production: true
 };